requirement.txt
Django>=5.0
djangorestframework
mysqlclient
drf-yasg
django-cors-headers
